+++ src/js/utils/helpers/BezierPathInteractionHelper.js
@@ -1,7 +1,6 @@
 "use strict";
 /**
- * @classdesc A helper for adding vertices to and remove vertices from Bézier paths.
- * By default the 'delete' key is used to remove vertices or paths.
+ * Interfaces and class for automatically handling Bézier curves.
  *
  * @requires AlloyFinger
  * @requires BezierPath
@@ -19,7 +18,9 @@
  * @date     2020-07-31
  * @modified 2020-08-03 Ported this class from vanilla JS to Typescript.
  * @modified 2020-08-12 Added a distance check before handling the click/tap event.
- * @version  1.0.1
+ * @modified 2021-01-03 Changed property to `autoAdjustPaths` in the HandlerOptions interface (typo).
+ * @modified 2021-01-03 Added following new functions: `addPathVertexDragStartListeners`, `removePathVertexDragStartListeners`, `addPathVertexDragEndListeners` and `removePathVertexDragEndListeners`.
+ * @version  1.1.0
  *
  * @file BezierPathInteractionHelper
  * @public
@@ -33,6 +34,17 @@ var MouseHandler_1 = require("../../MouseHandler");
 var PlotBoilerplate_1 = require("../../PlotBoilerplate");
 var Vertex_1 = require("../../Vertex");
 ;
+/**
+ * @classdesc A helper for adding vertices to and remove vertices from Bézier paths.
+ * By default the 'delete' key is used to remove vertices or paths.
+ *
+ * For convenience this helper is capable of handling multiple paths which are kept
+ * in an array.
+ *
+ * [Demo](https://www.plotboilerplate.io/repo/demos/23-bezier-point-distance/ "Demo")
+ *
+ * @public
+ **/
 var BezierPathInteractionHelper = /** @class */ (function () {
     /**
      * Pre: all paths must have been added to the PlotBoilerplate's drawable buffer (use the add(Drawable) function).
@@ -60,13 +72,20 @@ var BezierPathInteractionHelper = /** @class */ (function () {
         options = options || {};
         this.pb = pb;
         this.paths = [];
-        this.onPointerMoved = options.onPointerMoved || function (i, a, b, t) { };
-        this.onVertexInserted = options.onVertexInserted || function (i, j, n, o) { };
-        this.onVerticesDeleted = options.onVerticesDeleted || function (i, r, n, o) { },
-            this.onPathRemoved = options.onPathRemoved || function (i, o) { };
-        this.autoAdjustPaths = typeof options.autoAdjustPaths == 'boolean' ? options.autoAdjustPaths : true;
-        this.allowPathRemoval = typeof options.allowPathRemoval == 'boolean' ? options.allowPathRemoval : true;
-        this.maxDetectDistance = typeof options.maxDetectDistance == 'number' ? options.maxDetectDistance : Number.MAX_VALUE;
+        this.onPointerMoved =
+            typeof options.onPointerMoved === 'function' ? options.onPointerMoved : (function (i, a, b, t) { });
+        this.onVertexInserted =
+            typeof options.onVertexInserted === 'function' ? options.onVertexInserted : (function (i, j, n, o) { });
+        this.onVerticesDeleted =
+            typeof options.onVerticesDeleted === 'function' ? options.onVerticesDeleted : (function (i, r, n, o) { });
+        this.onPathRemoved =
+            typeof options.onPathRemoved === 'function' ? options.onPathRemoved : (function (i, o) { });
+        this.autoAdjustPaths =
+            typeof options.autoAdjustPaths === 'boolean' ? options.autoAdjustPaths : true;
+        this.allowPathRemoval =
+            typeof options.allowPathRemoval === 'boolean' ? options.allowPathRemoval : true;
+        this.maxDetectDistance =
+            typeof options.maxDetectDistance === 'number' ? options.maxDetectDistance : Number.MAX_VALUE;
         this.mouseIsOver = false;
         this.currentPathIndex = -1;
         this.currentDistance = Number.MAX_VALUE;
@@ -272,8 +291,6 @@ var BezierPathInteractionHelper = /** @class */ (function () {
         BezierPathInteractionHelper.setPathAutoAdjust(newPath);
         this.paths[pathIndex] = newPath;
         this.pb.add(newPath);
-        // Replace path inside the array with the new path
-        // this._addDefaultPathListeners( newPath );
     };
     ;
     // +---------------------------------------------------------------------------------
@@ -284,7 +301,6 @@ var BezierPathInteractionHelper = /** @class */ (function () {
         this.currentB.set(point);
         this._updateMinDistance();
         // Always fire even if nothing visually changed?
-        // console.log( this.currentDistance, this.maxDetectDistance, this.mouseIsOver );
         if ((this.currentDistance <= this.maxDetectDistance && this.mouseIsOver)
             && this.pb.getDraggedElementCount() == 0) {
             this.onPointerMoved(this.currentPathIndex, this.currentA, this.currentB, this.currentT);
@@ -348,7 +364,6 @@ var BezierPathInteractionHelper = /** @class */ (function () {
             // Check if there is already a path point at the given split position
             var pathPoint = path.getPointAt(_self.currentT);
             var pointNear = _self.pb.getVertexNear(_self.pb.revertMousePosition(pathPoint.x, pathPoint.y), 6.0);
-            // console.log( "pathPoint", pathPoint, pointNear );
             if (pointNear) {
                 for (var i = 0; i < path.bezierCurves.length; i++) {
                     if (path.bezierCurves[i].startPoint.distance(pointNear) <= 6.0 || path.bezierCurves[i].endPoint.distance(pointNear) <= 6.0) {
@@ -464,6 +479,66 @@ var BezierPathInteractionHelper = /** @class */ (function () {
     };
     ;
     /**
+     * A helper function to add drag-start-listener to given vertices.
+     */
+    /* private static _addVertsDragStartListener( verts:Array<Vertex>, dragStartListener:VertListener ) : void {
+    for( var i in verts ) {
+        verts[i].addDragStartListener( dragStartListener );
+    }
+    }; */
+    /**
+     * A helper function to remove drag-start-listener to given vertices.
+     */
+    /* private static _removeVertsDragStartListener( verts:Array<Vertex>, dragEndListener:VertListener ) : void {
+    for( var i in verts ) {
+        verts[i].removeDragStartListener( dragStartListener );
+    }
+    }; */
+    /**
+     * A helper function to add drag-start listeners to all vertices of the given path.
+     *
+     * @static
+     * @method addPathVertexDragStartListeners
+     * @memberof BezierPathInteractionHelper
+     * @param {BezierPath} path - The Bézier path to add vertex listeners to.
+     * @param {function} vertexDragStartListener - The drag listeners to add to each path vertex.
+     * @return void
+     **/
+    BezierPathInteractionHelper.addPathVertexDragStartListeners = function (path, vertexDragStartListener) {
+        for (var i = 0; i < path.bezierCurves.length; i++) {
+            var curve = path.bezierCurves[i];
+            curve.startPoint.listeners.addDragStartListener(vertexDragStartListener);
+            curve.startControlPoint.listeners.addDragStartListener(vertexDragStartListener);
+            curve.endControlPoint.listeners.addDragStartListener(vertexDragStartListener);
+            if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
+                curve.endPoint.listeners.addDragStartListener(vertexDragStartListener);
+            // if( i+1 == path.bezierCurves.length && !path.adjustCircular )
+            // BezierPathIntractionHelper._addVertsDragListener( [curve.startPoint, curve.startControlPoint, curve.endPoint, curve.endControlPoint ], vertexDragListener );
+        }
+    };
+    ;
+    /**
+     * A helper function to remove drag-start listeners to all vertices of the given path.
+     *
+     * @static
+     * @method removePathVertexDragStartListeners
+     * @memberof BezierPathInteractionHelper
+     * @param {BezierPath} path - The Bézier path to remove vertex listeners from.
+     * @param {function} vertexDragListener - The drag listeners to remove from each path vertex.
+     * @return void
+     **/
+    BezierPathInteractionHelper.removePathVertexDragStartListeners = function (path, vertexDragStartListener) {
+        for (var i = 0; i < path.bezierCurves.length; i++) {
+            var curve = path.bezierCurves[i];
+            curve.startPoint.listeners.removeDragStartListener(vertexDragStartListener);
+            curve.startControlPoint.listeners.removeDragStartListener(vertexDragStartListener);
+            curve.endControlPoint.listeners.removeDragStartListener(vertexDragStartListener);
+            if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
+                curve.endPoint.listeners.removeDragStartListener(vertexDragStartListener);
+        }
+    };
+    ;
+    /**
      * A helper function to add drag listeners to all vertices of the given path.
      *
      * @static
@@ -481,6 +556,8 @@ var BezierPathInteractionHelper = /** @class */ (function () {
             curve.endControlPoint.listeners.addDragListener(vertexDragListener);
             if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
                 curve.endPoint.listeners.addDragListener(vertexDragListener);
+            // if( i+1 == path.bezierCurves.length && !path.adjustCircular )
+            // BezierPathIntractionHelper._addVertsDragListener( [curve.startPoint, curve.startControlPoint, curve.endPoint, curve.endControlPoint ], vertexDragListener );
         }
     };
     ;
@@ -502,6 +579,50 @@ var BezierPathInteractionHelper = /** @class */ (function () {
             curve.endControlPoint.listeners.removeDragListener(vertexDragListener);
             if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
                 curve.endPoint.listeners.removeDragListener(vertexDragListener);
+        }
+    };
+    ;
+    /**
+     * A helper function to add drag-end listeners to all vertices of the given path.
+     *
+     * @static
+     * @method addPathVertexDragEndListeners
+     * @memberof BezierPathInteractionHelper
+     * @param {BezierPath} path - The Bézier path to add vertex listeners to.
+     * @param {function} vertexDragEndListener - The drag listeners to add to each path vertex.
+     * @return void
+     **/
+    BezierPathInteractionHelper.addPathVertexDragEndListeners = function (path, vertexDragEndListener) {
+        for (var i = 0; i < path.bezierCurves.length; i++) {
+            var curve = path.bezierCurves[i];
+            curve.startPoint.listeners.addDragEndListener(vertexDragEndListener);
+            curve.startControlPoint.listeners.addDragEndListener(vertexDragEndListener);
+            curve.endControlPoint.listeners.addDragEndListener(vertexDragEndListener);
+            if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
+                curve.endPoint.listeners.addDragEndListener(vertexDragEndListener);
+            // if( i+1 == path.bezierCurves.length && !path.adjustCircular )
+            // BezierPathIntractionHelper._addVertsDragListener( [curve.startPoint, curve.startControlPoint, curve.endPoint, curve.endControlPoint ], vertexDragListener );
+        }
+    };
+    ;
+    /**
+     * A helper function to remove drag-end listeners to all vertices of the given path.
+     *
+     * @static
+     * @method removePathVertexDragEndListeners
+     * @memberof BezierPathInteractionHelper
+     * @param {BezierPath} path - The Bézier path to remove vertex listeners from.
+     * @param {function} vertexDragListener - The drag listeners to remove from each path vertex.
+     * @return void
+     **/
+    BezierPathInteractionHelper.removePathVertexDragEndListeners = function (path, vertexDragEndListener) {
+        for (var i = 0; i < path.bezierCurves.length; i++) {
+            var curve = path.bezierCurves[i];
+            curve.startPoint.listeners.removeDragEndListener(vertexDragEndListener);
+            curve.startControlPoint.listeners.removeDragEndListener(vertexDragEndListener);
+            curve.endControlPoint.listeners.removeDragEndListener(vertexDragEndListener);
+            if (i + 1 == path.bezierCurves.length && !path.adjustCircular)
+                curve.endPoint.listeners.removeDragEndListener(vertexDragEndListener);
         }
     };
     ;